Move SinglePortfolioNetwork components into own file
